dev-requirements.txt
-r requirements.txt
responses==0.3.0
pytest-pep8==1.0.6
setuptools==11.3.1
sphinx==1.2.3
sphinx_rtd_theme==0.1.6
sphinxcontrib_napoleon==0.2.9
